Makefile: allow for combining DEVELOPER=1 and CFLAGS="..."
Ever since the DEVELOPER=1 facility introduced there's been no way to
set a custom set of CFLAGS on the command-line (e.g. CFLAGS="-O0 -g
-ggdb3") *and* benefit from the set of warnings and assertions
DEVELOPER=1 enables.

This is because the semantics of variables in the Makefile are such
that the user setting CFLAGS overrides anything we set, including what
we're doing in config.mak.dev[1].

So let's introduce a "DEVELOPER_CFLAGS" variable set by
config.mak.dev. We then add this to ALL_CFLAGS which, along with some
other logic in the Makefile, basically means our compiler flags are
now the equivalent of:

    ALL_CFLAGS = $(DEVELOPER_CFLAGS) $(CPPFLAGS) $(CFLAGS) $(BASIC_CFLAGS) $(EXTRA_CPPFLAGS)

The reason for putting DEVELOPER_CFLAGS is, as discussed in the docs
being added here, to allow for overriding something DEVELOPER=1 brings
in. On both GCC and Clang later settings override earlier
ones. E.g. "-Wextra -Wno-extra" will enable no "extra" warnings, but
not if those two arguments are reversed.

1. https://www.gnu.org/software/make/manual/html_node/Overriding.html

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>