clone: don't depend on fickle link(2) semantics
When we clone a repository locally without --no-hardlink we'll link()
those files we think we can hardlink.

But the semantics of link() are OS-specific. On GNU calling link() on
a symlink will produce another copy of that symlink, on
e.g. NetBSD (and seemingly OSX) we'll instead get a hardlink to the
file the symlink pointed to.

On Debian GNU/Linux:

    $ touch foo
    $ ln -s foo bar
    $ link bar zar
    $ file foo bar zar
    foo: empty
    bar: symbolic link to foo
    zar: symbolic link to foo

But on NetBSD:

    $ file foo bar zar
    foo: empty
    bar: symbolic link to `foo'
    zar: empty

However arbitrary these clone semantics are, let's not leave their
behavior to fickle OS-specific details. It's a fair bet that to the
extent anyone's cared about this it's been tested more on Linux, so
let's re-implement its behavior in a bug-compatible way for now.

1. https://public-inbox.org/git/CAHd-oW71N94vBvN4wYQVzXyUZjvVs3Ca9tM3VUEuqB66YNGDtA@mail.gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>