stash: setup default diff output format if necessary
In the scripted 'git stash show' when no arguments are passed, we just
pass '--stat' to 'git diff'.  When any argument is passed to 'stash
show', we no longer pass '--stat' to 'git diff', and pass whatever
flags are passed directly through to 'git diff'.

By default 'git diff' shows the patch output.  So when we a user uses
'git stash show -v', they would be shown the diff, because that's the
default behaviour of 'git diff', but not actually directly triggered
by passing the '-v'.

In the C version of 'git stash show', we try to emulate that
behaviour using the internal diff API.  However we forgot to set up
the default output format, in case it wasn't set by any of the flags
that were passed through.  So 'git stash show -v' in the builtin
version of stash would be completely silent, while it would show the
diff before.

Fix this by setting up the default output format for 'git diff'.

Note that "diff -v" is a fairly useless no-op, but options like "git
stash show --patience" shares the same issue that is fixed with this
change.

Reported-by: Denton Liu <liu.denton@gmail.com>
Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>