trace2: indicate hostname & PID and if PID wrapped
Change the SID format to <time>-H<host_hash>-P<pid>[WF] where W =
wrapped, F = Full.

On Linux we can have really big pids, but typically
/proc/sys/kernel/pid_max is left at the default of 32768, so 5 digits
should be enough for everyone.

But since this is also going to be used for ad-hoc debugging, let's
indicate in the field if the value is wrapped or not (so not the
actual pid). Before we'd just silently shorten it. This is mainly here
for the entropy, but also a little bit for ad-hoc debugging.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>