check-non-portable-shell: support Perl versions older than 5.10
For thoroughness when checking for one-shot environment variable
assignments at shell function call sites, check-non-portable-shell
stitches together incomplete lines (those ending with backslash). This
allows it to correctly flag such undesirable usage even when the
variable assignment and function call are split across lines, for
example:

    FOO=bar \
    func

where 'func' is a shell function.

The stitching is accomplished like this:

    while (<>) {
        chomp;
        # stitch together incomplete lines (those ending with "\")
        while (s/\\$//) {
            $_ .= readline;
            chomp;
        }
        # detect unportable/undesirable shell constructs
        ...
    }

Although this implementation is well supported in reasonably modern Perl
versions (5.10 and later), it fails in a couple ways with older versions
(such as Perl 5.8 shipped with ancient Mac OS 10.5).

In particular, in older Perl versions, 'readline' is not connected to
the file handle associated with the "magic" while (<>) {...} construct,
so 'readline' throws a "readline() on unopened filehandle" error.
Furthermore, $_ assigned by the outer while-loop is read-only, so the
attempt to modify it via "$_ .= readline" in the inner while-loop fails
with a "Modification of a read-only value" error.

Avoid both problems by collecting the stitched-together line in a
variable other than $_ and dropping the inner loop entirely.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>