fsck: test for 2^63 overflowing in addition to 2^64
Copy the test added in d4b8de0420 ("fsck: report integer overflow in
author timestamps", 2014-02-24) changing 18446744073709551617 to
9223372036854775809. This tests if a 64 bit signed timestamp_t and
time_t would overflow, in addition to the existing unsigned test.

The overflow check added in that change and amended in
7ca36d9398 ("date: check date overflow against time_t", 2014-02-24)
would already catch this, but it wasn't tested for.

On my system (and I believe all *nix) time_t is signed. POSIX just
says that it "shall be an integer type"[1].

1. https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_types.h.html

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>