fsck: report near-integer overflow in ident timestamps
Extend the checks for invalid timestamps to take into account that we
not only use the timestamp_t value we parse out of it, but add
whatever we find in the timezone offset to it.

Since we'll accept a timezone offset of +9999 we not only need to
check if the timestamp itself overflows, but whether it overflows when
99 hours and 99 minutes are added to it, since we'll shortly be adding
the parsed timezone offset to a variable stored in a timestamp_t.

Failing to exhaustively check for this means that a prankster can
craft a commit which fsck (and thus popular git hosting sites) will
accept, but that git itself will promptly die on when any command
which parses the commit object is run, such as "log", "show" etc.

This wasn't always the case. When the timestamp overflow check was
first added in d4b8de0420 ("fsck: report integer overflow in author
timestamps", 2014-02-24), such a timestamp would cause git to render a
bogus value of "Thu Jan 1 00:00:00 1970 +0000". That change was first
released as part of Git version 1.9.1.

Rendering a bogus timestamp is mildly annoying, but no big deal.

However, since 1e65a982da ("date.c: abort if the system time cannot
handle one of our timestamps", 2017-04-26) first released with Git
version 2.14.0 (on August 4th, 2017) a timestamp like this will cause
git to die on any command that tries to parse it. E.g.:

    $ ~/g/git/git --exec-path=$PWD show --no-abbrev --no-decorate; echo $?
    fatal: Timestamp too large for this system: 9223372036854783007
    commit 4cfe4cb4e202f64de8745f9f5ea64723939fb7cd
    128

See https://github.com/avar/Y292277026596 and
https://gitlab.com/avar/Y292277026596 for a repository containing such
a commit. Both of those hosting sites render the commit, but both
return an error of some sort if asked to produce a *.patch file, which
I suspect is because they're calling git-format-patch under the hood:
https://github.com/avar/Y292277026596/commit/4cfe4cb.patch &
https://gitlab.com/avar/Y292277026596/commit/4cfe4cb.patch

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>